refactoring of the VaultClient class is completed added decorators for re-authentication and notice about deprecated methods
