+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## v2.1.0 - 2024-06-03
+## v2.1.0 - 2024-06-04
 ### What's Changed
 **Full Changelog**: https://github.com/obervinov/vault-package/compare/v2.0.4...v2.1.0 by @obervinov in https://github.com/obervinov/vault-package/pull/47
 #### 💥 Breaking Changes
