+++ vault/__init__.py
@@ -7,11 +7,14 @@ from .client import VaultClient
 from .kv2_engine import KV2Engine
 from .db_engine import DBEngine
 from .exceptions import WrongKV2Configuration, WrongDBConfiguration
+from .decorators import reauthenticate_on_forbidden, deprecated_method
 
 __all__ = [
     'VaultClient',
     'KV2Engine',
     'DBEngine',
     'WrongKV2Configuration',
-    'WrongDBConfiguration'
+    'WrongDBConfiguration',
+    'reauthenticate_on_forbidden',
+    'deprecated_method'
 ]
