--- vault/deprecated.py
@@ -1,1 +1,0 @@
-"""This file generates warnings about methods and classes that will be removed in future versions"""
